fix: resolve project view rendering bugs
- Coerce non-string values in esc()/escAttr() to prevent crash when
  PostgreSQL returns Date objects for completed_at
- Replace separate onClipsChanged/onProjectsChanged with debounced
  scheduleReload() to prevent stale sidebar and project counts
- Cast COUNT(*)::int in getProjects query to fix PG bigint-as-string
- Fire projects-changed from delete-clip and assign-clip-to-project
- Open DevTools when SCIURUS_DEV=1 for easier debugging
- Add try/catch around renderClipCard in project detail view

+++ src/db-pg.js
@@ -321,7 +321,7 @@ async function deleteComment(commentId) {
 async function getProjects() {
   const { rows } = await pool.query(`
     SELECT p.*,
-           (SELECT COUNT(*) FROM clips WHERE project_id = p.id) AS "clipCount"
+           (SELECT COUNT(*) FROM clips WHERE project_id = p.id)::int AS "clipCount"
     FROM projects p
     ORDER BY p.name
   `);
